count_topologies tested
